Add better nix support, add GTK theme installation

+++ restore.sh
@@ -1,9 +1,19 @@
 #!/usr/bin/env bash
+shopt -s extglob
 GIT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)
+
+command_exists() { type "$1" &>/dev/null; }
 
 # fish
 rm -rf $HOME/.config/fish/
-cp -r $GIT_DIR/fish/ $HOME/.config/
+if command_exists "nixos-rebuild"; then
+    cd $GIT_DIR/fish/
+    mkdir $HOME/.config/fish/
+    cp -r !(config.fish) $HOME/.config/fish/
+    cd -
+else
+    cp -r $GIT_DIR/fish/ $HOME/.config/
+fi
 cp $GIT_DIR/universal-shellrc.txt $HOME/.config/fish/conf.d/universal.fish
 fish -c 'fisher update'
 
@@ -50,6 +60,21 @@ cp $GIT_DIR/claws-mail/matcherrc $HOME/.claws-mail/matcherrc
 # rofi
 rm -rf $HOME/.config/rofi/
 cp -r $GIT_DIR/rofi $HOME/.config/rofi/
+
+# Install catppuccin GTK theme
+# TODO: Add automatic catppuccin GTK updates - would be easy, but no version numbering once install, need to figure that out
+ctp_version="v0.7.1"
+variant="Catppuccin-Mocha-Standard-Mauve-Dark"
+if [ ! -e ~/.themes/$variant ] && [ ! command_exists "nixos-rebuild" ]; then
+    mkdir -p ~/.themes &&
+        curl -L https://github.com/catppuccin/gtk/releases/download/$ctp_version/$variant.zip -o ~/.themes/catppuccin.zip &&
+        unzip ~/.themes/catppuccin.zip -d ~/.themes/ &&
+        rm -rf ~/.themes/catppuccin.zip
+
+    # gtk 4
+    rm -rf $HOME/.config/gtk-4.0
+    ln -sf "${HOME}/.themes/${variant}/gtk-4.0/" "${HOME}/.config/gtk-4.0"
+fi
 
 # VS code
 mkdir -p $HOME/.config/Code/User/
